Add StitchBrokenDatasetCommand
This command works around bio-formats by parsing the positions in the nd file,
and then using IJ.open (via VisiviewUtils) to open the single stk files as tiffs.

The metadata (in particular the pixel spacing) are lost when reading as tiff,
so we have to ask for them in the input parameters.